AzBlob components: Extract shared code
